fix(api): Class rename
