+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/admin/AdminProfileChangeStatusResponseDTO.java
@@ -1,7 +1,7 @@
 package com.team18.FleetForge.dto.admin;
 
 
-import com.team18.FleetForge.model.enums.DriverProfileChangeRequestStatus;
+import com.team18.FleetForge.model.enums.InformationChangeRequestStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class AdminAcceptProfileChangeResponseDTO {
-    DriverProfileChangeRequestStatus status;
+public class AdminProfileChangeStatusResponseDTO {
+    InformationChangeRequestStatus status;
     Long driverId;
 }
