+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/driver/DriverProfileChangeResponseDTO.java
@@ -1,7 +1,7 @@
 package com.team18.FleetForge.dto.driver;
 
 import com.team18.FleetForge.model.DriverProfileChangeRequest;
-import com.team18.FleetForge.model.enums.DriverProfileChangeRequestStatus;
+import com.team18.FleetForge.model.enums.InformationChangeRequestStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DriverProfileChangeResponseDTO {
-    DriverProfileChangeRequestStatus status;
+    InformationChangeRequestStatus status;
     LocalDateTime createdAt;
     Long requestId;
