+++ backend/FleetForge/src/main/java/com/team18/FleetForge/dto/vehicle/VehicleCreateResponseDTO.java
@@ -1,6 +1,6 @@
 package com.team18.FleetForge.dto.vehicle;
 
-import com.team18.FleetForge.utils.VehicleType;
+import com.team18.FleetForge.model.enums.VehicleType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
