+++ backend/FleetForge/src/main/java/com/team18/FleetForge/model/Users/Admin.java
@@ -3,9 +3,11 @@ package com.team18.FleetForge.model.Users;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
+@SuperBuilder
 public class Admin extends User {
 }
