+++ backend/FleetForge/src/main/java/com/team18/FleetForge/model/Users/Passenger.java
@@ -1,10 +1,8 @@
 package com.team18.FleetForge.model.Users;
 
 import com.team18.FleetForge.model.Route;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
+import lombok.*;
+import lombok.experimental.SuperBuilder;
 
 import java.util.List;
 
@@ -12,6 +10,7 @@ import java.util.List;
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
+@SuperBuilder
 public class Passenger extends User {
     private List<Route> favoriteRoutes;
 }
