+++ backend/FleetForge/src/main/java/com/team18/FleetForge/model/Users/User.java
@@ -3,10 +3,12 @@ package com.team18.FleetForge.model.Users;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@SuperBuilder
 public abstract class User {
 
     private Long id;
