+++ backend/FleetForge/src/main/java/com/team18/FleetForge/model/Vehicle.java
@@ -1,6 +1,6 @@
 package com.team18.FleetForge.model;
 
-import com.team18.FleetForge.utils.VehicleType;
+import com.team18.FleetForge.model.enums.VehicleType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
