+++ backend/FleetForge/src/main/java/com/team18/FleetForge/model/enums/VehicleType.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.utils;
+package com.team18.FleetForge.model.enums;
 
 public enum VehicleType {
     STANDARD,
